requirements.txt
django
djangorestframework
beancount
python-dateutil
django-cors-headers
